+++ internal/vaultread/service_test.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/iasi777/v-memory/internal/checklist"
+	"github.com/iasi777/memauthority/internal/checklist"
 )
 
 func TestReadHandoffReturnsStructuredTodoChecklist(t *testing.T) {